Created a health component which, when attached, gives the player health and changes their state to dead when health reaches 0
