+++ Assets/Scripts/HealthComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7500cdb3c5976c8429cc49c83795c08e
+guid: c22bb4ddccf7c2d4e8fae6e7d1f08b23
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
